# backup_clean_code/docs-web/static/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for multi-tenant IncidentHub SOC platform.",
    "title": "IncidentHub API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/cases/{caseID}/communications/{threadID}/sync": {
      "post": {
        "operationId": "post_api_v1_cases_caseID_communications_threadID_sync",
        "parameters": [
          {
            "in": "path",
            "name": "caseID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "threadID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "POST /api/v1/cases/{caseID}/communications/{threadID}/sync"
      }
    },
...
    "/api/v1/forum/threads/{threadID}/proxy/sync": {
      "post": {
        "operationId": "post_api_v1_forum_threads_threadID_proxy_sync",
        "parameters": [
          {
            "in": "path",
            "name": "threadID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "POST /api/v1/forum/threads/{threadID}/proxy/sync"
      }
    },
...
    "/api/v1/operations/{operationID}": {
      "get": {
        "operationId": "get_api_v1_operations_operationID",
        "parameters": [
          {
            "in": "path",
            "name": "operationID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "GET /api/v1/operations/{operationID}"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ]
}